[Tests] Move is_inf tests to proper place.
