+++ src/tests/functional/shared_test_classes/include/shared_test_classes/single_layer/is_inf.hpp
